refactor: move RandomPrimeInRange() function to common package
