+++ internal/common/randomprime_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package gabi
+package common
 
 import (
 	"crypto/rand"
@@ -12,7 +12,7 @@ import (
 )
 
 func TestRandomPrimeInRange(t *testing.T) {
-	p, err := randomPrimeInRange(rand.Reader, 597, 120)
+	p, err := RandomPrimeInRange(rand.Reader, 597, 120)
 	assert.NoError(t, err)
 
 	assert.True(t, p.ProbablyPrime(22), "p not prime!")
